Split UI and API logic into separate modules for better maintainability. Made functions more modular to improve reusability and readability.
