+++ auth/index.html
@@ -103,19 +103,18 @@
           >
             Create an Account
           </a>
+          <!-- Divider -->
+          <div class="flex items-center my-6">
+            <div class="flex-1 border-t border-gray-300"></div>
+            <span class="mx-4 text-gray-500">or</span>
+            <div class="flex-1 border-t border-gray-300"></div>
+          </div>
           <a
             href="/auth/login/"
             class="block w-full bg-white text-blue-600 py-3 px-6 rounded-md text-center font-medium border border-blue-600 hover:bg-blue-50 transition-colors"
           >
             Log In
           </a>
-        </div>
-
-        <!-- Divider -->
-        <div class="flex items-center my-6">
-          <div class="flex-1 border-t border-gray-300"></div>
-          <span class="mx-4 text-gray-500">or</span>
-          <div class="flex-1 border-t border-gray-300"></div>
         </div>
 
         <!-- Additional Links -->
