+++ auth/login/index.html
@@ -89,6 +89,7 @@
                 type="email"
                 name="email"
                 id="email"
+                value="test1001@stud.noroff.no"
                 minlength="8"
                 required
                 placeholder="first.last@stud.noroff.no"
@@ -105,6 +106,7 @@
                 type="password"
                 name="password"
                 id="password"
+                value="passwordtest1001"
                 placeholder="Enter password"
                 required
                 class="mt-1 w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500"
