+++ src/js/router/views/login.js
@@ -1,5 +1,22 @@
 import { onLogin } from "../../ui/auth/login";
 
-const form = document.forms.login;
+/**
+ *
+ * Sets up an event listener for the login form.
+ *
+ * @function init
+ * @async
+ * @throws {Error} If there is an issue during initialization.
+ */
+async function init() {
+  try {
+    // Set up an event listener for the login form
+    const form = document.forms.login;
+    form.addEventListener("submit", onLogin);
+  } catch (error) {
+    console.error("Error during initialization:", error.message);
+  }
+}
 
-form.addEventListener("submit", onLogin);
+// Initialize the application
+init();
